% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq_27.R
\name{.get_hc}
\alias{.get_hc}
\title{THIS DOCUMENTATION IS COPIED FROM ANOTHER FUNCTION. USE IT AS A TEMPLATE TO MODIFY FOR NEW FUNCTIONS BELOW
P_a: density of dry air (g / m^3)}
\usage{
.get_hc(pars, constants)
}
\value{
Value in g / m\eqn{^3} of class \code{units}
}
\description{
THIS DOCUMENTATION IS COPIED FROM ANOTHER FUNCTION. USE IT AS A TEMPLATE TO MODIFY FOR NEW FUNCTIONS BELOW
P_a: density of dry air (g / m^3)
}
\details{
\deqn{P_\mathrm{a} = P / (R_\mathrm{air} (T_\mathrm{leaf} - T_\mathrm{air}) / 2)}{P_a = P / (R_air (T_leaf - T_air) / 2)}

\tabular{lllll}{
\emph{Symbol} \tab \emph{R} \tab \emph{Description} \tab \emph{Units} \tab \emph{Default}\cr
\eqn{P} \tab \code{P} \tab atmospheric pressure \tab kPa \tab 101.3246\cr
\eqn{R_\mathrm{air}}{R_air} \tab \code{R_air} \tab specific gas constant for dry air \tab J / (kg K) \tab 287.058\cr
\eqn{T_\mathrm{air}}{T_air} \tab \code{T_air} \tab air temperature \tab K \tab 298.15\cr
\eqn{T_\mathrm{leaf}}{T_leaf} \tab \code{T_leaf} \tab leaf temperature \tab K \tab input
}
}
\examples{

library(tealeaves)

cs <- make_constants()
ep <- make_enviropar()
lp <- make_leafpar()

T_leaf <- set_units(298.15, K)

tealeaves:::.get_Pa(T_leaf, c(cs, ep, lp), FALSE)

}
